Add functionality to download analysis on startup
If the `sarif-viewer.githubCodeScanningInitialAlert` setting has a
value, use the contents of the value as the alert to apply. This is
meant as a way to automatically apply fixes from a codescanning alert.

Additionally, if this alert has any fixes in it, they are automatically
applied. We also need to handle some slight differences in the SARIF
format. The fix objects that we are receiving are non-standard. We
expect that for each fix, there is a `diff` in its property bag that
contains a git-patch formatted version of the diff to apply.

Lastly, this adds some simple telemetry. It collects that the
autofix was applied, whether or not it was successful, and (if
unsuccessful) a failure message.